--- packages/twenty-server/src/engine/metadata-modules/field-metadata/interceptors/field-metadata-graphql-api-exception.interceptor.ts
@@ -1,34 +1,0 @@
-import {
-  type CallHandler,
-  type ExecutionContext,
-  Injectable,
-  type NestInterceptor,
-} from '@nestjs/common';
-import { GqlExecutionContext } from '@nestjs/graphql';
-
-import { type Observable, catchError } from 'rxjs';
-import { SOURCE_LOCALE } from 'twenty-shared/translations';
-
-import { I18nService } from 'src/engine/core-modules/i18n/i18n.service';
-import { fieldMetadataGraphqlApiExceptionHandler } from 'src/engine/metadata-modules/field-metadata/utils/field-metadata-graphql-api-exception-handler.util';
-
-@Injectable()
-export class FieldMetadataGraphqlApiExceptionInterceptor
-  implements NestInterceptor
-{
-  constructor(private readonly i18nService: I18nService) {}
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
-    const gqlContext = GqlExecutionContext.create(context);
-    const ctx = gqlContext.getContext();
-    const locale = ctx.req?.locale ?? SOURCE_LOCALE;
-    const i18n = this.i18nService.getI18nInstance(locale);
-
-    return next
-      .handle()
-      .pipe(
-        catchError((err) => fieldMetadataGraphqlApiExceptionHandler(err, i18n)),
-      );
-  }
-}
