--- packages/twenty-server/src/engine/metadata-modules/field-metadata/utils/is-field-metadata-type-morph-relation.util.ts
@@ -1,10 +1,0 @@
-import { FieldMetadataType } from 'twenty-shared/types';
-
-import { type FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
-
-export const isFieldMetadataTypeMorphRelation = (
-  fieldMetadata: FieldMetadataEntity,
-): fieldMetadata is FieldMetadataEntity &
-  FieldMetadataEntity<FieldMetadataType.MORPH_RELATION> => {
-  return fieldMetadata.type === FieldMetadataType.MORPH_RELATION;
-};
