--- packages/twenty-server/src/engine/metadata-modules/field-metadata/utils/is-select-field-metadata-type.util.ts
@@ -1,6 +1,0 @@
-import { FieldMetadataType } from 'twenty-shared/types';
-export const isSelectFieldMetadataType = (
-  type: FieldMetadataType,
-): type is FieldMetadataType.SELECT => {
-  return type === FieldMetadataType.SELECT;
-};
