--- packages/twenty-server/src/engine/metadata-modules/object-metadata/utils/build-name-label-and-description-for-foreign-key-field-metadata.util.ts
@@ -1,15 +1,0 @@
-export const buildNameLabelAndDescriptionForForeignKeyFieldMetadata = ({
-  targetObjectNameSingular,
-  targetObjectLabelSingular,
-  relatedObjectLabelSingular,
-}: {
-  targetObjectNameSingular: string;
-  targetObjectLabelSingular: string;
-  relatedObjectLabelSingular: string;
-}) => {
-  const name = `${targetObjectNameSingular}Id`;
-  const label = `${targetObjectLabelSingular} ID (foreign key)`;
-  const description = `${relatedObjectLabelSingular} ${targetObjectLabelSingular} id foreign key`;
-
-  return { name, label, description };
-};
