--- packages/twenty-server/src/engine/metadata-modules/utils/is-arg-defined-if-provided-or-throw.util.ts
@@ -1,15 +1,0 @@
-import { isDefined } from 'twenty-shared/utils';
-export const isArgDefinedIfProvidedOrThrow = ({
-  input,
-  key,
-  value,
-}: {
-  input: object;
-  key: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  value: any;
-}) => {
-  if (key in input && !isDefined(value)) {
-    throw new Error(`${key} must be defined when provided`);
-  }
-};
