--- packages/twenty-server/src/engine/metadata-modules/utils/validate-metadata-name-is-camel-case.utils.ts
@@ -1,16 +1,0 @@
-import { t } from '@lingui/core/macro';
-import camelCase from 'lodash.camelcase';
-
-import {
-  InvalidMetadataException,
-  InvalidMetadataExceptionCode,
-} from 'src/engine/metadata-modules/utils/exceptions/invalid-metadata.exception';
-
-export const validateMetadataNameIsCamelCaseOrThrow = (name: string) => {
-  if (name !== camelCase(name)) {
-    throw new InvalidMetadataException(
-      t`${name} should be in camelCase`,
-      InvalidMetadataExceptionCode.NOT_CAMEL_CASE,
-    );
-  }
-};
